+++ packages/g-shopify-library/src/settings/index.ts
@@ -37,30 +37,51 @@ export type SettingsMapped<T extends Record<string, Setting>> = Array<
             : never
     }[keyof T]
 >
-export const single_setting_schema = z.union([
-    z.intersection(
-        z.discriminatedUnion('type', [
-            BasicSettings.text,
-            BasicSettings.textarea,
-            BasicSettings.number,
-            BasicSettings.radio,
-            BasicSettings.checkbox,
-            BasicSettings.select,
-            BasicSettings.range,
-            ...Object.values(SpecializedSettings),
-            ...Object.values(ShopifySettings),
-        ]),
-        baseSchema
-    ),
-    z.discriminatedUnion('type', [
-        SidebarSettings.paragraph,
-        SidebarSettings.header,
-    ]),
-])
-export const setting_schema = z.array(single_setting_schema)
 
-export type SingleSetting = z.infer<typeof single_setting_schema>
-export type Settings = z.infer<typeof setting_schema>
+export const getSettingSchema = (id: string | undefined = undefined) => {
+    let baseSchemaID = z.object({
+        label: z.string().optional(), ///todo: not sure if for shopify
+        info: z.string().optional(),
+        id: z.string(),
+    })
+
+    if (id !== undefined) {
+        baseSchemaID = z.object({
+            label: z.string().optional(), ///todo: not sure if for shopify
+            info: z.string().optional(),
+            id: z.string().regex(new RegExp(id)),
+        })
+    }
+
+    return z.union([
+        z.intersection(
+            z.discriminatedUnion('type', [
+                BasicSettings.text,
+                BasicSettings.textarea,
+                BasicSettings.number,
+                BasicSettings.radio,
+                BasicSettings.checkbox,
+                BasicSettings.select,
+                BasicSettings.range,
+                ...Object.values(SpecializedSettings),
+                ...Object.values(ShopifySettings),
+            ]),
+            baseSchemaID
+        ),
+
+        z.discriminatedUnion('type', [
+            SidebarSettings.paragraph,
+            SidebarSettings.header,
+        ]),
+    ])
+}
+export const getSettingGroupSchema = () => {
+    return z.array(getSettingSchema())
+}
+
+export type SingleSetting = z.infer<ReturnType<typeof getSettingSchema>>
+export type Settings = z.infer<ReturnType<typeof getSettingGroupSchema>>
+export type SettingGroup = Settings
 
 export const parseValidatorFactory =
     <T extends z.ZodTypeAny>(schema: T) =>
@@ -68,7 +89,11 @@ export const parseValidatorFactory =
         return schema.safeParse(data).success
     }
 const getSettingsFactory =
-    <T extends typeof setting_schema | typeof single_setting_schema>(
+    <
+        T extends
+            | ReturnType<typeof getSettingGroupSchema>
+            | ReturnType<typeof getSettingSchema>
+    >(
         schema: T
     ) =>
     (data: unknown): z.infer<T> | undefined => {
@@ -78,8 +103,45 @@ const getSettingsFactory =
         return undefined
     }
 
-export const parseSettings = getSettingsFactory(setting_schema)
-export const parseSingleSetting = getSettingsFactory(single_setting_schema)
+export const parseSettings = getSettingsFactory(getSettingGroupSchema())
+
+export const parseSettingsGroup = (
+    data: unknown,
+    id_prefix: string | undefined = undefined,
+    id_suffix: string | undefined = undefined
+) => {
+    const _schema = getSettingGroupSchema().transform((data_array) => {
+        /* * Transform result to add prefix/suffix * */
+        return data_array.map((data_value) => {
+            let _data_value = data_value
+            if (
+                data_value.type !== 'header' &&
+                data_value.type !== 'paragraph'
+            ) {
+                _data_value = {
+                    ...data_value,
+                    id: `${id_prefix !== undefined ? id_prefix : ''}${
+                        data_value.id
+                    }${id_suffix !== undefined ? id_suffix : ''}`,
+                }
+            }
+            return _data_value
+        })
+    })
+    if (parseValidatorFactory(_schema)(data)) {
+        return _schema.parse(data)
+    } else return undefined
+}
+
+export const parseSingleSetting = (
+    data: unknown,
+    id: string | undefined = undefined
+) => {
+    const _schema = getSettingSchema(id)
+    if (parseValidatorFactory(_schema)(data)) {
+        return _schema.parse(data)
+    } else return undefined
+}
 export const parseSetting = parseSingleSetting
 
 export {}